remove more old code and move some things to helper
